moved register contract logic to scenario crate
