+++ framework/scenario/Cargo.toml
@@ -29,6 +29,10 @@ colored = "2.0"
 [features]
 run-go-tests = []
 
+[dependencies.multiversx-sc-meta]
+version = "=0.39.5"
+path = "../meta"
+
 [dependencies.multiversx-chain-vm]
 version = "=0.1.5"
 path = "../../vm"
